feat: update select on storybook
