+++ dashboard/components/select/Select.mocks.tsx
@@ -1,10 +1,11 @@
 import { SelectProps } from './Select';
 
 const base: SelectProps = {
+  value: 'o1',
   label: 'Select',
   values: ['o1', 'o2', 'o3'],
   displayValues: ['Option 1', 'Option 2', 'Option 3'],
-  onChange: () => {}
+  handleChange: () => {}
 };
 
 const mockSelectProps = {
